requirements.txt
python-telegram-bot==13.13
requests==2.28
firebase-admin==5.2
load-dotenv==0.1
